@@ -6,7 +6,7 @@ namespace serve
     public class Startup
     {
 #pragma warning disable CA1822 // Mark members as static
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             var fileServerOptions = new FileServerOptions {
                 EnableDefaultFiles = true,
